+++ tests/test_detail.py
@@ -1,0 +1,48 @@
+import re
+
+import pytest
+from flask import url_for
+
+from registry.donor.models import Note
+
+from .fixtures import sample_of_rc
+from .helpers import login
+
+
+class TestDetail:
+    @pytest.mark.parametrize("rodne_cislo", sample_of_rc(50))
+    def test_detail(self, user, testapp, rodne_cislo):
+        """Just a simple test that the detail page loads for some random donors"""
+        login(user, testapp)
+        res = testapp.get(url_for("donor.detail", rc=rodne_cislo))
+        assert res.status_code == 200
+        assert "<td></td>" not in res
+        # Check that the sum of the donations is eqal to the total count
+        donations_list = re.search(r"<h2>Počty darování</h2>(.*?)</ul>", res.text, re.S)
+        numbers = re.findall(r">.*?: (\d+)</", donations_list.group())
+        numbers = list(map(int, numbers))
+        assert sum(numbers[:-1]) == numbers[-1]
+
+    @pytest.mark.parametrize("rodne_cislo", sample_of_rc(5))
+    def test_save_update_note(self, user, testapp, rodne_cislo):
+        existing_notes = Note.query.count()
+        login(user, testapp)
+        res = testapp.get(url_for("donor.detail", rc=rodne_cislo))
+        # New note
+        form = res.forms["noteForm"]
+        assert form.fields["note"][0].value == ""
+        form.fields["note"][0].value = "Lorem ipsum"
+        res = form.submit().follow()
+        assert res.status_code == 200
+        assert "Poznámka uložena." in res
+        assert "Lorem ipsum</textarea>" in res.text
+        assert Note.query.count() == existing_notes + 1
+        # Update existing
+        form = res.forms["noteForm"]
+        assert form.fields["note"][0].value == "Lorem ipsum"
+        form.fields["note"][0].value += " dolor sit amet,"
+        res = form.submit().follow()
+        assert res.status_code == 200
+        assert "Poznámka uložena." in res
+        assert "Lorem ipsum dolor sit amet,</textarea>" in res.text
+        assert Note.query.count() == existing_notes + 1
